+++ Domain.Sql.Tests/Domain.Sql.Tests.csproj
@@ -145,7 +145,7 @@
     <Compile Include="ReadModelCatchupRollingCatchupTests.cs" />
     <Compile Include="SqlCommandSchedulerTests.cs" />
     <Compile Include="SqlCommandSchedulerTests_Legacy.cs" />
-    <Compile Include="SqlCommandSchedulerTests_New.cs" />
+    <Compile Include="SqlCommandSchedulerTests_Pipeline.cs" />
     <Compile Include="SqlEventSourcedRepositoryMigrationTests.cs" />
     <Compile Include="SqlEventSourcedRepositoryTests.cs" />
     <Compile Include="StorableEventTests.cs" />
